fix(potentiometer): emit value in `detail` property
Currently, we emit an `input` even whenever the potentiometer value changes. This is on par with how `<input type="range">` works.

However, the InputEvent emitted by `range` does not include any `data` property, and MDN says that "data" should include only the inserted characters, if any:

https://developer.mozilla.org/en-US/docs/Web/API/InputEvent/data

Thus, we use the `detail` property instead, which is also consistent with what we do in the membrane-keypad element.